Conflicts:
	Dockerfile
	README.md
	start.sh

+++ start.sh
@@ -6,6 +6,13 @@ backup=$(ls -p -t backup*.tar.gz | head -1)
 
 : ${PRIV_KEY_NAME=priv.key}
 
-sshpass -p "$PASS" scp -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no  -i /certs/$PRIV_KEY_NAME  $backup $DISTANT_HOST_PATH/$backup
+if [ "$TYPE"  = "SCP" ]
+ then sshpass -p "$PASS" scp -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no  -i /certs/$PRIV_KEY_NAME  $backup $DISTANT_HOST_PATH/$backup
+fi
 
+if [ "$TYPE"  = "FTP" ]
+ then  lftp ftp://$DISTANT_HOST_PATH -e "mirror -Re /save $BACKUP_PATH;exit"
+fi
+
+#Should be keep to syncro all  and remove only oldest
 rm $backup
